feat(cmd): split roles into dedicated subcommands
